activity_filter/pyproject.toml
[project]
name = "activity-filter"
version = "0.1.0"
description = ""
authors = [
    {name = "yujhenchen",email = "yujhen.chen93@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pymongo[srv] (>=4.12.0,<5.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "activity_filter", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
